# cmd/docs/openapiv3.json
{"components":{"requestBodies":{"CreateCARequest":{"content":{"application/json":{"schema":{"properties":{"ca_ttl":{"type":"integer"},"enroller_ttl":{"type":"integer"},"key_metadata":{"properties":{"bits":{"type":"integer"},"type":{"type":"string"}},"type":"object"},"subject":{"$ref":"#/components/schemas/CertificateSubject"}}}}},"description":"Request used for creating a new Certificate Authority","required":true},"ImportCARequest":{"content":{"application/json":{"schema":{"properties":{"ca_ttl":{"type":"integer"},"enroller_ttl":{"type":"integer"},"key_metadata":{"properties":{"bits":{"type":"integer"},"type":{"type":"string"}},"type":"object"},"subject":{"$ref":"#/components/schemas/CertificateSubject"}}}}},"description":"Request used for importing a Certificate Authority","required":true},"SignCertificateRequest":{"content":{"application/json":{"schema":{"properties":{"csr":{"type":"string"}}}}},"description":"Request used for getting all a certificates","required":true}},"responses":{"CreateCAResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LamassuCA"}}},"description":"Response returned back after creating a CA."},"DeleteCertResponse":{"content":{"application/json":{"schema":{}}},"description":"Response returned back after importing a CA."},"ErrorResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"}}}}},"description":"Response when errors happen."},"GetCAResponse":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/LamassuCA"},"type":"array"}}},"description":"Response returned back after getting a CA."},"GetCertResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LamassuCA"}}},"description":"Response returned back after getting a certificate."},"GetIssuedCertsResponse":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/LamassuCA"},"type":"array"}}},"description":"Response returned back after getting list of issued certificates."},"HealthResponse":{"content":{"application/json":{"schema":{"properties":{"healthy":{"type":"boolean"}}}}},"description":"Response returned back after healthchecking."},"ImportCAResponse":{"content":{"application/json":{"schema":{}}},"description":"Response returned back after importing a CA."},"RevokeCAResponse":{"content":{"application/json":{"schema":{}}},"description":"Response returned back after revoking a CA."},"SignCertificateResponse":{"content":{"application/json":{"schema":{"properties":{"crt":{"type":"boolean"}}}}},"description":"Response returned back after signing certificate."}},"schemas":{"CAType":{"enum":["pki","dmsenroller"],"type":"string"},"CertificateContent":{"properties":{"pem_base64":{"type":"string"},"public_key_base64":{"type":"string"}},"type":"object"},"CertificateStatus":{"enum":["issued","expired"],"type":"string"},"CertificateSubject":{"properties":{"common_name":{"type":"string"},"country":{"type":"string"},"locality":{"type":"string"},"organization":{"type":"string"},"organization_unit":{"type":"string"},"state":{"type":"string"}},"type":"object"},"KeyMetadata":{"properties":{"bits":{"type":"integer"},"strength":{"type":"string"},"type":{"type":"string"}},"type":"object"},"LamassuCA":{"properties":{"ca_ttl":{"type":"integer"},"certificate":{"$ref":"#/components/schemas/CertificateContent"},"enroller_ttl":{"type":"integer"},"key_metadata":{"$ref":"#/components/schemas/KeyMetadata"},"name":{"type":"string"},"serial_number":{"type":"string"},"status":{"$ref":"#/components/schemas/CertificateStatus"},"subject":{"$ref":"#/components/schemas/CertificateSubject"},"valid_from":{"type":"string"},"valid_to":{"type":"string"}},"type":"object"}}},"info":{"contact":{"url":"https://github.com/lamassuiot"},"description":"REST API used for interacting with Lamassu CA","license":{"name":"MPL v2.0","url":"https://github.com/lamassuiot/lamassu-compose/blob/main/LICENSE"},"title":"Lamassu CA API","version":"0.0.0"},"openapi":"3.0.0","paths":{"/v1/health":{"get":{"description":"Get health status","operationId":"Health","responses":{"200":{"$ref":"#/components/responses/HealthResponse"}}}},"/v1/pki/import/{caName}":{"post":{"description":"Import existing crt and key","operationId":"ImportCA","parameters":[{"in":"path","name":"caName","required":true,"schema":{"type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/ImportCARequest"},"responses":{"200":{"$ref":"#/components/responses/ImportCAResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/pki/{caName}":{"delete":{"description":"Revoke CA","operationId":"RevokeCA","parameters":[{"in":"path","name":"caName","required":true,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/RevokeCAResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"post":{"description":"Create new CA using Form","operationId":"CreateCA","parameters":[{"in":"path","name":"caName","required":true,"schema":{"type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/CreateCARequest"},"responses":{"200":{"$ref":"#/components/responses/CreateCAResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/{caType}":{"get":{"description":"Get all CAs","operationId":"GetCAs","parameters":[{"in":"path","name":"caType","required":true,"schema":{"enum":["pki","dmsenroller"]}}],"responses":{"200":{"$ref":"#/components/responses/GetCAResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/{caType}/{caName}/cert/{serialNumber}":{"delete":{"description":"Revoke certificate issued by {caName} and {serialNumber}","operationId":"DeleteCert","parameters":[{"in":"path","name":"caName","required":true,"schema":{"type":"string"}},{"in":"path","name":"caType","required":true,"schema":{"enum":["pki","dmsenroller"]}},{"in":"path","name":"serialNumber","required":true,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/DeleteCertResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"get":{"description":"Get certificate by {caName} and {serialNumber}","operationId":"GetCert","parameters":[{"in":"path","name":"caName","required":true,"schema":{"type":"string"}},{"in":"path","name":"caType","required":true,"schema":{"enum":["pki","dmsenroller"]}},{"in":"path","name":"serialNumber","required":true,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/GetCertResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/{caType}/{caName}/issued":{"get":{"description":"Get Issued certificates by {caName}","operationId":"GetIssuedCerts","parameters":[{"in":"path","name":"caName","required":true,"schema":{"type":"string"}},{"in":"path","name":"caType","required":true,"schema":{"enum":["pki","dmsenroller"]}}],"responses":{"200":{"$ref":"#/components/responses/GetIssuedCertsResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/v1/{caType}/{caName}/sign":{"post":{"description":"Sign CSR by {caName}","operationId":"SignCertificate","parameters":[{"in":"path","name":"caName","required":true,"schema":{"type":"string"}},{"in":"path","name":"caType","required":true,"schema":{"enum":["pki","dmsenroller"]}}],"requestBody":{"$ref":"#/components/requestBodies/SignCertificateRequest"},"responses":{"200":{"$ref":"#/components/responses/SignCertificateResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"401":{"$ref":"#/components/responses/ErrorResponse"},"403":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}}},"servers":[{"description":"Current Server","url":"/"}]}